refactor(webcam): remove adapter; turn it into a hook
