+++ bigbluebutton-html5/imports/ui/components/app/container.jsx
@@ -22,11 +22,11 @@ import useUserChangedLocalSettings from '../../services/settings/hooks/useUserCh
 import { PINNED_PAD_SUBSCRIPTION } from '../notes/queries';
 import connectionStatus from '../../core/graphql/singletons/connectionStatus';
 import useDeduplicatedSubscription from '../../core/hooks/useDeduplicatedSubscription';
-import VideoStreamsState from '../video-provider/state';
 import useSettings from '../../services/settings/hooks/useSettings';
 import { SETTINGS } from '../../services/settings/enums';
 import { useStorageKey } from '../../services/storage/hooks';
 import useMuteMicrophone from '../audio/audio-graphql/hooks/useMuteMicrophone';
+import { useVideoStreamsCount } from '../video-provider/hooks';
 
 const currentUserEmoji = (currentUser) => (currentUser
   ? {
@@ -128,7 +128,7 @@ const AppContainer = (props) => {
   const isSharedNotesPinnedFromGraphql = !!pinnedPadData
     && pinnedPadData.sharedNotes[0]?.sharedNotesExtId === NOTES_CONFIG.id;
   const isSharedNotesPinned = sharedNotesInput?.isPinned && isSharedNotesPinnedFromGraphql;
-  const isThereWebcam = VideoStreamsState.getStreams().length > 0;
+  const isThereWebcam = useVideoStreamsCount() > 0;
   const muteMicrophone = useMuteMicrophone();
   const isScreenSharingEnabled = useIsScreenSharingEnabled();
   const isExternalVideoEnabled = useIsExternalVideoEnabled();
