+++ bigbluebutton-html5/imports/ui/components/video-provider/types.ts
@@ -74,3 +74,4 @@ export type Stream = {
 export type StreamItem = Stream | ConnectingStream;
 export type GridItem = GridUser & { type: typeof VIDEO_TYPES.GRID };
 export type VideoItem = StreamItem | GridItem;
+export type StreamSubscriptionData = VideoStreamsResponse['user_camera'][number];
